Apply review feedback: fix circular import, move business logic, add select_for_update, extract fixtures to conftest, regenerate codegen
- Move inline `Game` import in tasks.py to module level via `apps.get_model()`
- Use `%s`-style logging instead of f-strings
- Remove no-op `validate()` from MemberSerializer; move confirmation deadline logic into `create()`
- Simplify `MemberCreateView.perform_create()` to delegate entirely to serializer
- Wrap `MemberConfirmView.post()` in `transaction.atomic()` with `select_for_update()` to prevent race conditions
- Extract all 12 test fixtures from test_staging_experience.py into game/tests/conftest.py
- Regenerate OpenAPI schema and TypeScript client; add `confirmed`, `confirmationRequired`, `confirmationDeadline` to mock fixtures
